NearBeachAPI/requirements.txt
#Requirements for NearBeach API
djangorestframework
django-oauth-toolkit